Refactor and add access code to forms
