+++ php/utils/assert_user_is_logged_in.php
@@ -1,0 +1,8 @@
+<?php
+session_start();
+
+if (!isset($_SESSION['user_faculty_number'])) {
+    header("Location: index.php");
+    exit;
+}
+?>
